Some additional problems added in function

+++ function_question/number_print.cpp
@@ -3,11 +3,11 @@ using namespace std;
 
 int printNum(int digit[]) {
 
-
     int ans=0;
-    for(int i=0; i < sizeof(digit); i++) {
+    for(size_t i=0; i< sizeof(digit); i++) {
         ans= ans*10+digit[i];
     }
+
     return ans;
 }
